Reorder the type parameters so that Scala does not bug out.
The bug is reported as https://github.com/scala/bug/issues/11845.

+++ src/main/scala/ru/ifmo/onell/problem/LinearRandomDoubleWeights.scala
@@ -7,7 +7,7 @@ import ru.ifmo.onell.util.{Helpers, OrderedSet}
 
 class LinearRandomDoubleWeights(val problemSize: Int, val maxWeight: Double, randomSeed: Long)
   extends HasEvaluation[Array[Boolean], Double]
-    with HasIncrementalEvaluation[Array[Boolean], Int, Double]
+    with HasIncrementalEvaluation[Array[Boolean], Double, Int]
 {
   private[this] val rng = new Random(randomSeed)
   private[this] val weights = Array.fill(problemSize)(rng.nextDouble() * (maxWeight - 1) + 1)
